progress++

+++ app/src/main/java/com/example/world_bank_v4/CaricaDati.java
@@ -10,20 +10,22 @@ import android.view.MenuItem;
 import android.view.View;
 import android.widget.AdapterView;
 import android.widget.ArrayAdapter;
+import android.widget.Button;
 import android.widget.CursorAdapter;
+import android.widget.ImageButton;
 import android.widget.ListView;
 
 import java.util.ArrayList;
 
 
 /*activity che carica e visualizza le n-tuple salvate nel database*/
-public class CaricaDati extends AppCompatActivity
-        implements AdapterView.OnItemClickListener, View.OnClickListener {
+public class CaricaDati extends AppCompatActivity implements View.OnClickListener {
 
 
     private DbManager dbManager;
     private ListView listView;
     private CursorAdapter cursorAdapter;
+
 
 
 
@@ -39,8 +41,10 @@ public class CaricaDati extends AppCompatActivity
         getSupportActionBar().setDisplayUseLogoEnabled(true);
 
         listView = findViewById(R.id.list_view_carica_dati);
-        dbManager = new DbManager(this);
-        new CaricaDatabaseTask(dbManager).execute();
+
+
+
+        new CaricaDatabaseTask().execute();
 
     }
 
@@ -105,18 +109,16 @@ public class CaricaDati extends AppCompatActivity
 
 
     @Override
-    public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
-
-    }
-
-
-
-    @Override
     public void onClick(View v) {
-        int position = listView.getPositionForView(v);
-        long id = cursorAdapter.getItemId(position);
-        if (dbManager.delete(id))
-            cursorAdapter.changeCursor(dbManager.query());
+        if(v.getId() == R.id.imageButtonDelete){
+            int position = listView.getPositionForView(v);
+            long id = cursorAdapter.getItemId(position);
+            if (dbManager.delete(id))
+            /*Change the underlying cursor to a new cursor. If there is an existing cursor it will
+             be closed*/
+                cursorAdapter.changeCursor(dbManager.query());
+        }
+
     }
 
 
@@ -129,16 +131,19 @@ public class CaricaDati extends AppCompatActivity
         private DbManager dbManager;
 
 
-        public CaricaDatabaseTask(DbManager dbManager){
-            this.dbManager = dbManager;
+        public CaricaDatabaseTask(){
+
+            this.dbManager = new DbManager(getApplicationContext()); /*oggetto per interagire con il
+                                                                database*/
+            setDbManager(dbManager);        /*passa alla classe contenitore un riferimento al
+                                            dbManager così lo potrà chiudere nella onDestroy()*/
         }
 
 
         @Override
         protected Cursor doInBackground(Void... params) {
 
-            dbManager = new DbManager(getApplicationContext()); /*oggetto per interagire con il
-                                                                database*/
+
             Cursor cursor = dbManager.query();
 
             return cursor;
@@ -149,9 +154,9 @@ public class CaricaDati extends AppCompatActivity
         protected void onPostExecute(Cursor cursorRisultato){
             Log.d(Costanti.NOME_APP , ": CURSORE -->  "+ showCursor(cursorRisultato));
 
-            cursorAdapter = new MyCursorAdapter(getApplicationContext(), cursorRisultato, 0,
-                    getThis());
-            listView.setAdapter(cursorAdapter);
+            caricaLayout(cursorRisultato);
+
+
         }
 
 
@@ -171,5 +176,29 @@ public class CaricaDati extends AppCompatActivity
     }
 
 
+
+
+
+    private void caricaLayout(Cursor cursorRisultato){
+
+
+        cursorAdapter = new MyCursorAdapter(this, cursorRisultato, 0,
+                getThis());
+
+        listView.setAdapter(cursorAdapter);
+
+
+
+
+
+    }
+
+
+
     public CaricaDati getThis() {return  this;}
+
+    public void setDbManager(DbManager dbManager){
+        this.dbManager = dbManager;
+    }
+
 }
